Extract functionality into Environment class
Environment and RailsEnvironment can enforce the
presence of the keys listed in the yaml file.
